requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-jose
passlib
bcrypt
pymongo